requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
python-dotenv==0.20.0
reportlab==3.6.12
pillow==9.2.0